+++ common/infinitealloys/tile/TEMEnergyStorage.java
@@ -98,7 +98,7 @@ public class TEMEnergyStorage extends TileEntityMachine {
 			coords.add((short)point.y);
 			coords.add(point.z);
 		}
-		return ArrayUtils.addAll(super.getSyncDataToClient(), connectedMachines.size(), coords.toArray());
+		return ArrayUtils.addAll(super.getSyncDataToClient(), (byte)connectedMachines.size(), coords.toArray());
 	}
 
 	public void handlePacketDataFromServer(int currentRK) {
